feat: convert general ledger accounts-list modal to a page
