+++ react-app/src/pages/Client/Accounting/DisbursementsPage.jsx
@@ -1,4 +1,3 @@
-import { truncate } from "lodash";
 import Layout from "../../../components/Layouts/client/Layout.jsx";
 import useDisbursements from "../../../hooks/page-hooks/useDisbursements.js";
 import DisbursementsAsyncSelect from "../../../components/DisbursementsAsyncSelect.jsx";
